[refactor] 인터랙션 섹션 폴더 경로 변경
