+++ src/mainPage/features/interactions/description/IntroductionDetail.jsx
@@ -1,4 +1,4 @@
-import EventDescriptionLayout from "@/eventDescription/EventDescriptionLayout.jsx";
+import EventDescriptionLayout from "@main/eventDescription/EventDescriptionLayout.jsx";
 
 export default function IntroductionDetail({ prizes }) {
   return (
