+++ src/mainPage/features/interactions/description/TapBar.jsx
